Data alignment issues, continued
Trying to abstract the alignment required by atomic types for all complex objects so the class users don’t have to care about it.
Currently alignment is defaulted to 16 everywhere, but it’s not perfect because LockFreeQueue stores a dummy cell internally, and the content of cells may require a greater alignment. Unfortunately alignof() is not supported by VS2013 at the moment.